Create Item.java helper class and mention in README
Map deals with a pair of data - a key and an associated value.
Together they will be referred to as an "Item". This type will be
used in future Map implementations to support returning both the
key and its assicated value.

The README file contains a directory tree that was was updated to
reflect the change in the working tree.

The implementation and documentation of Item.java is now complete.